# openapi/a2a_v0.54.0.yaml
openapi: 3.1.0
info:
  title: Loop Controller A2A Protocol
  version: 0.54.0
  description: |
    Current v0.54.0 HTTP/JSON wire authority. AgentSpec/AgentStatus, scheduling,
    assignment claim/renew/result and execution attempts are internal capabilities;
    no public scheduler or worker endpoint exists. Assignment observation is exposed
    only through task snapshots and dead-letter operations. AgentCard.version is the
    agent implementation version; protocol_version is the negotiated wire version.
servers:
  - url: http://127.0.0.1:8080
security: []
paths:
...
  /a2a/v1/delegation-approvals/{id}:
    get:
      operationId: getDelegationApproval
      security: [{ControlBearer: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '200': {description: Approval, content: {application/json: {schema: {$ref: '#/components/schemas/DelegationApproval'}}}}
        '401': {$ref: '#/components/responses/Error401'}
        '403': {$ref: '#/components/responses/Error403'}
        '404': {$ref: '#/components/responses/Error404'}
  /a2a/v1/delegation-approvals/{id}/approve:
    post:
      operationId: approveDelegation
      security: [{ApproverBearer: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      requestBody: {$ref: '#/components/requestBodies/ApprovalAction'}
      responses:
        '200': {description: Consumed approval, content: {application/json: {schema: {$ref: '#/components/schemas/DelegationApproval'}}}}
        '400': {$ref: '#/components/responses/Error400'}
        '401': {$ref: '#/components/responses/Error401'}
        '403': {$ref: '#/components/responses/Error403'}
        '409': {$ref: '#/components/responses/Error409'}
  /a2a/v1/delegation-approvals/{id}/reject:
    post:
      operationId: rejectDelegation
      security: [{ApproverBearer: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      requestBody: {$ref: '#/components/requestBodies/ApprovalAction'}
      responses:
        '200': {description: Rejected approval, content: {application/json: {schema: {$ref: '#/components/schemas/DelegationApproval'}}}}
        '400': {$ref: '#/components/responses/Error400'}
        '401': {$ref: '#/components/responses/Error401'}
        '403': {$ref: '#/components/responses/Error403'}
        '409': {$ref: '#/components/responses/Error409'}
  /a2a/v1/delegation-approvals/{id}/cancel:
    post:
      operationId: cancelDelegationApproval
      security: [{ControlBearer: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      requestBody: {$ref: '#/components/requestBodies/ApprovalAction'}
      responses:
        '200': {description: Pending/approved returns DelegationApproval; consumed returns Task, content: {application/json: {schema: {oneOf: [{$ref: '#/components/schemas/DelegationApproval'}, {$ref: '#/components/schemas/Task'}]}}}}
        '400': {$ref: '#/components/responses/Error400'}
        '401': {$ref: '#/components/responses/Error401'}
        '403': {$ref: '#/components/responses/Error403'}
        '409': {$ref: '#/components/responses/Error409'}
...
components:
  securitySchemes:
    ControlBearer: {type: http, scheme: bearer, description: Control principal; optional only in explicitly configured local compatibility mode.}
    ApproverBearer: {type: http, scheme: bearer, description: Independent approval principal.}
    TaskDelegationToken: {type: http, scheme: bearer, bearerFormat: JWT, description: Task-scoped delegation token.}
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string}}
...
  requestBodies:
    ApprovalAction: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ApprovalActionRequest'}}}}
  responses:
    Error400: {description: Bad request, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    Error401: {description: Authentication required, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    Error403: {description: Access denied, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    Error404: {description: Not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    Error409: {description: Conflict, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
  schemas:
    ProtocolVersion: {type: string, const: 0.54.0}
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [protocol_version, error, code]
      properties: {protocol_version: {$ref: '#/components/schemas/ProtocolVersion'}, error: {type: string}, code: {type: string}}
...
    Budget:
      type: object
      additionalProperties: false
      properties: {token_count: {type: integer}, payment_amount: {type: number}, currency: {type: string}}
...
    Task:
      type: object
      additionalProperties: false
      required: [protocol_version, task_id, session_id, initiator_agent_id, target_agent_id, status, created_at, updated_at, allow_redelegation]
      properties: {protocol_version: {$ref: '#/components/schemas/ProtocolVersion'}, task_id: {type: string}, request_id: {type: string}, session_id: {type: string}, interaction_id: {type: string}, decision_id: {type: string}, root_interaction_id: {type: string}, parent_interaction_id: {type: string}, root_task_id: {type: string}, parent_task_id: {type: string}, delegation_depth: {type: integer}, deadline: {type: string, format: date-time}, budget: {$ref: '#/components/schemas/Budget'}, reserved_budget: {$ref: '#/components/schemas/Budget'}, consumed_budget: {$ref: '#/components/schemas/Budget'}, allowed_tools: {type: array, items: {type: string}}, allowed_capabilities: {type: array, items: {type: string}}, allow_redelegation: {type: boolean}, initiator_agent_id: {type: string}, target_agent_id: {type: string}, status: {type: string, enum: [pending, accepted, running, completed, failed, cancelled, outcome_unknown]}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}, completed_at: {type: string, format: date-time}, outcome: {}, error_code: {type: string}, tenant_id: {type: string}, target_workload_id: {type: string}, target_instance_id: {type: string}}
...
    DelegationApproval:
      type: object
      additionalProperties: false
      required: [protocol_version, approval_id, request_id, decision_id, request_hash, initiator_agent_id, target_agent_id, session_id, delegation_depth, allowed_tools, allowed_capabilities, allow_redelegation, budget, expires_at, status, created_at, updated_at, version]
      properties: {protocol_version: {$ref: '#/components/schemas/ProtocolVersion'}, approval_id: {type: string}, request_id: {type: string}, decision_id: {type: string}, request_hash: {type: string}, initiator_agent_id: {type: string}, target_agent_id: {type: string}, session_id: {type: string}, tenant_id: {type: string}, target_workload_id: {type: string}, target_instance_id: {type: string}, root_task_id: {type: string}, parent_task_id: {type: string}, delegation_depth: {type: integer}, allowed_tools: {type: array, items: {type: string}}, allowed_capabilities: {type: array, items: {type: string}}, allow_redelegation: {type: boolean}, budget: {$ref: '#/components/schemas/Budget'}, task_deadline: {type: string, format: date-time}, expires_at: {type: string, format: date-time}, status: {type: string}, approver_id: {type: string}, reason: {type: string}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}, decided_at: {type: string, format: date-time}, task_id: {type: string}, version: {type: integer}}
    ApprovalActionRequest:
      type: object
      additionalProperties: false
      required: [request_id]
      properties: {request_id: {type: string}, reason: {type: string}}
...